Fix CGALConfig.cmake generation
The cmake files for this project require CMAKE_INSTALL_LIBDIR to be
appended to CMAKE_INSTALL_PREFIX, so override the default value to
remove the prefix (i.e. "lib64" instead of "/usr/lib64").

This fixes the generation of CGALConfig.cmake and a few other cmake
files, thereby fixing the build for SFCGAL.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>